Working tabs, CourseTitle, and workaround for editing highlighting + commented ugly try code

+++ src/components/course-editor/editable-item.js
@@ -1,9 +1,10 @@
 import React, {useState} from 'react'
 import {NavLink, Link, useLocation, useRouteMatch} from "react-router-dom";
-import {act} from "@testing-library/react";
 
 const EditableItem = (
     {
+        // setEditForParent,
+        // childEditing,
         path,
         deleteItem,
         updateItem,
@@ -14,44 +15,52 @@ const EditableItem = (
     const [editing, setEditing] = useState(false)
     // CachedItem is to save the current item to change title while editing
     const [cachedItem, setCachedItem] = useState(item)
-    const [selected, setSelected] = useState(active)
+    // const [selected, setSelected] = useState(active)
     // const location = useLocation()
     // const location = useRouteMatch(path)
 
     return (
         <>
             { !editing &&
-            <>
-                <Link className={`nav-link d-inline-block text-truncate col-8 ${active ? 'active' : ''}`}
-                      to={path}>
-                    {item.title} {JSON.stringify(active)} {item._id}
-                </Link>
-                <span className="col-3 float-right text-right pt-2">
-                    <i onClick={() => setEditing(true)}
-                       className="fas fa-edit fa-lg"/>
-                </span>
-            </>
+                <>
+                    <Link className={`btn text-left text-truncate d-inline-block col-12 ${active ? "active text-white" : ""}`}
+                          to={path}>
+                        {item.title}
+                        <i onClick={(event) =>
+                        {
+                            event.preventDefault()
+                            setEditing(true)
+                        }}
+                           className="fas fa-edit fa-lg float-right pt-1"/>
+                    </Link>
+                </>
             }
             { editing &&
-            <>
-                <input className="col-8 form-control d-inline-block"
+                <div className="col-12 bg-primary btn text-left">
+                <input className="form-control col-7 custom-control-inline"
                    onChange={(event) => setCachedItem(
                        {...cachedItem, title: event.target.value}
                    )}
-                   value={cachedItem.title}/>
-                <span className="col-3 float-right text-right pt-2">
+                   value={cachedItem.title}>
+                </input>
+                <span className="float-right pt-2">
                     <i onClick={() => {
                         setEditing(false)
                         updateItem(cachedItem)
                     }}
-                       className="fas fa-check fa-lg mr-3 text-success"/>
+                       className={`fas fa-check fa-lg mr-3 text-white`}/>
                     <i onClick={() => {
-                        setEditing(false)
-                        deleteItem(cachedItem)
+                        if (!active) {
+                            setEditing(false)
+                            deleteItem(cachedItem)
+                        }
+                        else{
+                            alert(`Invalid delete: ${item.title}`)
+                        }
                     }}
-                       className="fas fa-minus-circle fa-lg text-danger"/>
+                       className={`fas fa-minus-circle fa-lg text-white`}/>
                 </span>
-            </>
+                </div>
             }
         </>
     )
